refactor: route app runtime facades
