+++ rustfs/src/storage/mod.rs
@@ -77,7 +77,9 @@ pub(crate) mod ecstore_capacity {
 }
 
 pub(crate) mod ecstore_client {
-    pub(crate) use rustfs_ecstore::api::client::{admin_handler_utils, object_api_utils, transition_api};
+    #[cfg(test)]
+    pub(crate) use rustfs_ecstore::api::client::transition_api;
+    pub(crate) use rustfs_ecstore::api::client::{admin_handler_utils, object_api_utils};
 }
 
 pub(crate) mod ecstore_compression {
@@ -188,7 +190,9 @@ pub(crate) mod ecstore_storage {
 
 pub(crate) mod ecstore_tier {
     pub(crate) use rustfs_ecstore::api::tier::tier::TierConfigMgr;
-    pub(crate) use rustfs_ecstore::api::tier::{tier, tier_admin, tier_config, tier_handlers, warm_backend};
+    #[cfg(test)]
+    pub(crate) use rustfs_ecstore::api::tier::warm_backend;
+    pub(crate) use rustfs_ecstore::api::tier::{tier, tier_admin, tier_config, tier_handlers};
 }
 
 pub(crate) const BUCKET_ACCELERATE_CONFIG: &str = ecstore_bucket::metadata::BUCKET_ACCELERATE_CONFIG;
